Modify the core Peridigm class and Field manager for correct functioning of CrackTip_Location compute class
A GLOBAL, SCALAR field called TimeStep_Count is defined in Field.h. Additionally, CrackTip_Location field is
added to Field.h.

TimeStep_Count is a counter for the number of time steps in an explicit time integration simulation. It is
required for successful execution of CrackTip_Location compute class. Peridigm.cpp/.hpp files are modified to
incorporate the TimeStep_Count variable.

Changes to be committed:
	modified:   src/core/Peridigm.cpp
	modified:   src/core/Peridigm.hpp
	modified:   src/io/mesh_output/Field.h